DataUtils

DataUtils

Useful utilities and tools

The source repository is https://github.com/jlapeyre/DataUtils.jl.

No names are exported.

If some of this code duplicates functionality available elsewhere, or if it would find a good home in a well-supported package, please open an issue at the github repository.

Contents

DataUtils

Contents

Index

Functions

Index

DataUtils.countmap

DataUtils.writecopyconstructor

Functions

DataUtils.countmap — Function.

countmap(iter; datatype=Any, dicttype=Dict)

Return an AbstractDict counting the number of occurences of elements in iter.

Unlike StatsBase.countmap, DataUtils.countmap does not require that one first calls collect on iter if iter is not an AbstractArray. DataUtils.countmap accepts any iterator, including one that may generate more values than can be stored in memory.

DataUtils.countmap does not implement a fast, radix sort as StatsBase.countmap does.

Example

julia> import StatsBase; import DataUtils;

julia> d = Dict(rand(1:1000) => rand(1:5) for i in 1:100);

julia> StatsBase.countmap(values(d))
ERROR: MethodError: no method matching countmap(::Base.ValueIterator{Dict{Int64,Int64}})
...

julia> DataUtils.countmap(values(d))
Dict{Any,Int64} with 5 entries:
  4 => 19
  2 => 21
  3 => 20
  5 => 11
  1 => 25

source

countmap(iter::AbstractArray{T}; dicttype=Dict) where {T}

Equivalent to calling countmap(iter; datatype=Any, dicttype=Dict), with datatype=T.

source

DataUtils.writecopyconstructor — Function.

writecopyconstructor(mod, T; evaluate=true)

Constructs and compiles a simple copy constructor for struct type T. Each field is copied if a corresponding method for copy exists, and is passed unaltered otherwise. The code creating the copy constructor is evaluated in module mod. If evaluate is false, the expression defining the copy method is returned and is not evaluted.

Note that the copy methods for the fields must exist at the time writecopyconstructor is called.

Example

struct A
 a::Int
 b::String
 c::Vector
end

julia> DataUtils.writecopyconstructor(Main, A)

julia> x = A(1, "hi", [1, 2])
A(1, "hi", [1, 2])

julia> y = copy(x)
A(1, "hi", [1, 2])

julia> x.c[1] = 7;

julia> x
A(1, "hi", [7, 2])

julia> y
A(1, "hi", [1, 2])

source